Factory address route param (#454)
* support factory address as route param

* replace 'projects' route name with 'round'

* make login message a function

* move app logic to factory view
