+++ vue-app/src/api/user.ts
@@ -3,17 +3,8 @@ import { BigNumber, Contract } from 'ethers'
 import { Web3Provider } from '@ethersproject/providers'
 
 import { UserRegistry, ERC20 } from './abi'
-import { factory, ipfsGatewayUrl, provider } from './core'
+import { ipfsGatewayUrl, provider } from './core'
 import { BrightId } from './bright-id'
-
-//TODO: update anywhere this is called to take factory address as a parameter, default to env. variable
-export const LOGIN_MESSAGE = `Welcome to clr.fund!
-
-To get logged in, sign this message to prove you have access to this wallet. This does not cost any ether.
-
-You will be asked to sign each time you load the app.
-
-Contract address: ${factory.address.toLowerCase()}.`
 
 export interface User {
   walletAddress: string
@@ -62,4 +53,14 @@ export async function getEtherBalance(
   walletAddress: string
 ): Promise<BigNumber> {
   return await provider.getBalance(walletAddress)
+}
+
+export function getLoginMessage(factoryAddress: string) {
+  return `Welcome to clr.fund!
+
+To get logged in, sign this message to prove you have access to this wallet. This does not cost any ether.
+
+You will be asked to sign each time you load the app.
+
+Contract address: ${factoryAddress.toLowerCase()}.`
 }
