+++ vue-app/src/store/mutation-types.ts
@@ -4,6 +4,7 @@ export const SET_CURRENT_USER = 'SET_CURRENT_USER'
 export const SET_CURRENT_ROUND_ADDRESS = 'SET_CURRENT_ROUND_ADDRESS'
 export const SET_CURRENT_ROUND = 'SET_CURRENT_ROUND'
 export const SET_ACTIVE_ROUND_ADDRESS = 'SET_ACTIVE_ROUND_ADDRESS'
+export const SET_CURRENT_FACTORY_ADDRESS = 'SET_CURRENT_FACTORY_ADDRESS'
 export const SET_TALLY = 'SET_TALLY'
 export const SET_CONTRIBUTOR = 'SET_CONTRIBUTOR'
 export const SET_CONTRIBUTION = 'SET_CONTRIBUTION'
