+++ vue-app/src/views/ProjectAdded.vue
@@ -28,7 +28,7 @@
             </li>
           </ul>
           <div class="mt2 button-spacing">
-            <links to="/projects" class="btn-primary">View projects</links>
+            <links to="/round" class="btn-primary">View projects</links>
             <links to="/" class="btn-secondary">Go home</links>
           </div>
         </div>
